Enum aws_sdk_sfn::Error

#[non_exhaustive]
pub enum Error {

    ActivityDoesNotExist(ActivityDoesNotExist),
    ActivityLimitExceeded(ActivityLimitExceeded),
    ActivityWorkerLimitExceeded(ActivityWorkerLimitExceeded),
    ExecutionAlreadyExists(ExecutionAlreadyExists),
    ExecutionDoesNotExist(ExecutionDoesNotExist),
    ExecutionLimitExceeded(ExecutionLimitExceeded),
    InvalidArn(InvalidArn),
    InvalidDefinition(InvalidDefinition),
    InvalidExecutionInput(InvalidExecutionInput),
    InvalidLoggingConfiguration(InvalidLoggingConfiguration),
    InvalidName(InvalidName),
    InvalidOutput(InvalidOutput),
    InvalidToken(InvalidToken),
    InvalidTracingConfiguration(InvalidTracingConfiguration),
    MissingRequiredParameter(MissingRequiredParameter),
    ResourceNotFound(ResourceNotFound),
    StateMachineAlreadyExists(StateMachineAlreadyExists),
    StateMachineDeleting(StateMachineDeleting),
    StateMachineDoesNotExist(StateMachineDoesNotExist),
    StateMachineLimitExceeded(StateMachineLimitExceeded),
    StateMachineTypeNotSupported(StateMachineTypeNotSupported),
    TaskDoesNotExist(TaskDoesNotExist),
    TaskTimedOut(TaskTimedOut),
    TooManyTags(TooManyTags),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ActivityDoesNotExist(ActivityDoesNotExist)

The specified activity does not exist.

Tuple Fields of ActivityDoesNotExist

0: ActivityDoesNotExist

ActivityLimitExceeded(ActivityLimitExceeded)

The maximum number of activities has been reached. Existing activities must be deleted before a new activity can be created.

Tuple Fields of ActivityLimitExceeded

0: ActivityLimitExceeded

ActivityWorkerLimitExceeded(ActivityWorkerLimitExceeded)

The maximum number of workers concurrently polling for activity tasks has been reached.

Tuple Fields of ActivityWorkerLimitExceeded

0: ActivityWorkerLimitExceeded

ExecutionAlreadyExists(ExecutionAlreadyExists)

The execution has the same name as another execution (but a different input).

Executions with the same name and input are considered idempotent.

Tuple Fields of ExecutionAlreadyExists

0: ExecutionAlreadyExists

ExecutionDoesNotExist(ExecutionDoesNotExist)

The specified execution does not exist.

Tuple Fields of ExecutionDoesNotExist

0: ExecutionDoesNotExist

ExecutionLimitExceeded(ExecutionLimitExceeded)

The maximum number of running executions has been reached. Running executions must end or be stopped before a new execution can be started.

Tuple Fields of ExecutionLimitExceeded

0: ExecutionLimitExceeded

InvalidArn(InvalidArn)

The provided Amazon Resource Name (ARN) is invalid.

Tuple Fields of InvalidArn

0: InvalidArn

InvalidDefinition(InvalidDefinition)

The provided Amazon States Language definition is invalid.

Tuple Fields of InvalidDefinition

0: InvalidDefinition

InvalidExecutionInput(InvalidExecutionInput)

The provided JSON input data is invalid.

Tuple Fields of InvalidExecutionInput

0: InvalidExecutionInput

InvalidLoggingConfiguration(InvalidLoggingConfiguration)

Tuple Fields of InvalidLoggingConfiguration

0: InvalidLoggingConfiguration

InvalidName(InvalidName)

The provided name is invalid.

Tuple Fields of InvalidName

0: InvalidName

InvalidOutput(InvalidOutput)

The provided JSON output data is invalid.

Tuple Fields of InvalidOutput

0: InvalidOutput

InvalidToken(InvalidToken)

The provided token is invalid.

Tuple Fields of InvalidToken

0: InvalidToken

InvalidTracingConfiguration(InvalidTracingConfiguration)

Your tracingConfiguration key does not match, or enabled has not been set to true or false.

Tuple Fields of InvalidTracingConfiguration

0: InvalidTracingConfiguration

MissingRequiredParameter(MissingRequiredParameter)

Request is missing a required parameter. This error occurs if both definition and roleArn are not specified.

Tuple Fields of MissingRequiredParameter

0: MissingRequiredParameter

ResourceNotFound(ResourceNotFound)

Could not find the referenced resource. Only state machine and activity ARNs are supported.

Tuple Fields of ResourceNotFound

0: ResourceNotFound

StateMachineAlreadyExists(StateMachineAlreadyExists)

A state machine with the same name but a different definition or role ARN already exists.

Tuple Fields of StateMachineAlreadyExists

0: StateMachineAlreadyExists

StateMachineDeleting(StateMachineDeleting)

The specified state machine is being deleted.

Tuple Fields of StateMachineDeleting

0: StateMachineDeleting

StateMachineDoesNotExist(StateMachineDoesNotExist)

The specified state machine does not exist.

Tuple Fields of StateMachineDoesNotExist

0: StateMachineDoesNotExist

StateMachineLimitExceeded(StateMachineLimitExceeded)

The maximum number of state machines has been reached. Existing state machines must be deleted before a new state machine can be created.

Tuple Fields of StateMachineLimitExceeded

0: StateMachineLimitExceeded

StateMachineTypeNotSupported(StateMachineTypeNotSupported)

Tuple Fields of StateMachineTypeNotSupported

0: StateMachineTypeNotSupported

TaskDoesNotExist(TaskDoesNotExist)

Tuple Fields of TaskDoesNotExist

0: TaskDoesNotExist

TaskTimedOut(TaskTimedOut)

Tuple Fields of TaskTimedOut

0: TaskTimedOut

TooManyTags(TooManyTags)

You've exceeded the number of tags allowed for a resource. See the Limits Topic in the AWS Step Functions Developer Guide.

Tuple Fields of TooManyTags

0: TooManyTags

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>