Module aws_sdk_sfn::error

Modules

activity_does_not_exist

See ActivityDoesNotExist

activity_limit_exceeded

See ActivityLimitExceeded

activity_worker_limit_exceeded

See ActivityWorkerLimitExceeded

execution_already_exists

See ExecutionAlreadyExists

execution_does_not_exist

See ExecutionDoesNotExist

execution_limit_exceeded

See ExecutionLimitExceeded

invalid_arn

See InvalidArn

invalid_definition

See InvalidDefinition

invalid_execution_input

See InvalidExecutionInput

invalid_logging_configuration

See InvalidLoggingConfiguration

invalid_name

See InvalidName

invalid_output

See InvalidOutput

invalid_token

See InvalidToken

invalid_tracing_configuration

See InvalidTracingConfiguration

missing_required_parameter

See MissingRequiredParameter

resource_not_found

See ResourceNotFound

state_machine_already_exists

See StateMachineAlreadyExists

state_machine_deleting

See StateMachineDeleting

state_machine_does_not_exist

See StateMachineDoesNotExist

state_machine_limit_exceeded

See StateMachineLimitExceeded

state_machine_type_not_supported

See StateMachineTypeNotSupported

task_does_not_exist

See TaskDoesNotExist

task_timed_out

See TaskTimedOut

too_many_tags

See TooManyTags

Structs

ActivityDoesNotExist

The specified activity does not exist.

ActivityLimitExceeded

The maximum number of activities has been reached. Existing activities must be deleted

ActivityWorkerLimitExceeded

The maximum number of workers concurrently polling for activity tasks has been

CreateActivityError

CreateStateMachineError

DeleteActivityError

DeleteStateMachineError

DescribeActivityError

DescribeExecutionError

DescribeStateMachineError

DescribeStateMachineForExecutionError

ExecutionAlreadyExists

The execution has the same name as another execution (but a different

ExecutionDoesNotExist

The specified execution does not exist.

ExecutionLimitExceeded

The maximum number of running executions has been reached. Running executions must end or

GetActivityTaskError

GetExecutionHistoryError

InvalidArn

The provided Amazon Resource Name (ARN) is invalid.

InvalidDefinition

The provided Amazon States Language definition is invalid.

InvalidExecutionInput

The provided JSON input data is invalid.

InvalidLoggingConfiguration

InvalidName

The provided name is invalid.

InvalidOutput

The provided JSON output data is invalid.

InvalidToken

The provided token is invalid.

InvalidTracingConfiguration

Your tracingConfiguration key does not match, or enabled has not been set to true or false.

ListActivitiesError

ListExecutionsError

ListStateMachinesError

ListTagsForResourceError

MissingRequiredParameter

Request is missing a required parameter. This error occurs if both definition

ResourceNotFound

Could not find the referenced resource. Only state machine and activity ARNs are

SendTaskFailureError

SendTaskHeartbeatError

SendTaskSuccessError

StartExecutionError

StartSyncExecutionError

StateMachineAlreadyExists

A state machine with the same name but a different definition or role ARN already

StateMachineDeleting

The specified state machine is being deleted.

StateMachineDoesNotExist

The specified state machine does not exist.

StateMachineLimitExceeded

The maximum number of state machines has been reached. Existing state machines must be

StateMachineTypeNotSupported

StopExecutionError

TagResourceError

TaskDoesNotExist

TaskTimedOut

TooManyTags

You've exceeded the number of tags allowed for a resource. See the Limits Topic in the

UntagResourceError

UpdateStateMachineError

Enums

CreateActivityErrorKind

CreateStateMachineErrorKind

DeleteActivityErrorKind

DeleteStateMachineErrorKind

DescribeActivityErrorKind

DescribeExecutionErrorKind

DescribeStateMachineErrorKind

DescribeStateMachineForExecutionErrorKind

GetActivityTaskErrorKind

GetExecutionHistoryErrorKind

ListActivitiesErrorKind

ListExecutionsErrorKind

ListStateMachinesErrorKind

ListTagsForResourceErrorKind

SendTaskFailureErrorKind

SendTaskHeartbeatErrorKind

SendTaskSuccessErrorKind

StartExecutionErrorKind

StartSyncExecutionErrorKind

StopExecutionErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateStateMachineErrorKind