Module aws_sdk_sfn::model

Modules

activity_failed_event_details

See ActivityFailedEventDetails

activity_list_item

See ActivityListItem

activity_schedule_failed_event_details

See ActivityScheduleFailedEventDetails

activity_scheduled_event_details

See ActivityScheduledEventDetails

activity_started_event_details

See ActivityStartedEventDetails

activity_succeeded_event_details

See ActivitySucceededEventDetails

activity_timed_out_event_details

See ActivityTimedOutEventDetails

billing_details

See BillingDetails

cloud_watch_events_execution_data_details

See CloudWatchEventsExecutionDataDetails

cloud_watch_logs_log_group

See CloudWatchLogsLogGroup

execution_aborted_event_details

See ExecutionAbortedEventDetails

execution_failed_event_details

See ExecutionFailedEventDetails

execution_list_item

See ExecutionListItem

execution_started_event_details

See ExecutionStartedEventDetails

execution_succeeded_event_details

See ExecutionSucceededEventDetails

execution_timed_out_event_details

See ExecutionTimedOutEventDetails

history_event

See HistoryEvent

history_event_execution_data_details

See HistoryEventExecutionDataDetails

lambda_function_failed_event_details

See LambdaFunctionFailedEventDetails

lambda_function_schedule_failed_event_details

See LambdaFunctionScheduleFailedEventDetails

lambda_function_scheduled_event_details

See LambdaFunctionScheduledEventDetails

lambda_function_start_failed_event_details

See LambdaFunctionStartFailedEventDetails

lambda_function_succeeded_event_details

See LambdaFunctionSucceededEventDetails

lambda_function_timed_out_event_details

See LambdaFunctionTimedOutEventDetails

log_destination

See LogDestination

logging_configuration

See LoggingConfiguration

map_iteration_event_details

See MapIterationEventDetails

map_state_started_event_details

See MapStateStartedEventDetails

state_entered_event_details

See StateEnteredEventDetails

state_exited_event_details

See StateExitedEventDetails

state_machine_list_item

See StateMachineListItem

tag

See Tag

task_failed_event_details

See TaskFailedEventDetails

task_scheduled_event_details

See TaskScheduledEventDetails

task_start_failed_event_details

See TaskStartFailedEventDetails

task_started_event_details

See TaskStartedEventDetails

task_submit_failed_event_details

See TaskSubmitFailedEventDetails

task_submitted_event_details

See TaskSubmittedEventDetails

task_succeeded_event_details

See TaskSucceededEventDetails

task_timed_out_event_details

See TaskTimedOutEventDetails

tracing_configuration

See TracingConfiguration

Structs

ActivityFailedEventDetails

Contains details about an activity that failed during an execution.

ActivityListItem

Contains details about an activity.

ActivityScheduleFailedEventDetails

Contains details about an activity schedule failure that occurred during an

ActivityScheduledEventDetails

Contains details about an activity scheduled during an execution.

ActivityStartedEventDetails

Contains details about the start of an activity during an execution.

ActivitySucceededEventDetails

Contains details about an activity that successfully terminated during an

ActivityTimedOutEventDetails

Contains details about an activity timeout that occurred during an execution.

BillingDetails

An object that describes workflow billing details.

CloudWatchEventsExecutionDataDetails

Provides details about execution input or output.

CloudWatchLogsLogGroup

ExecutionAbortedEventDetails

Contains details about an abort of an execution.

ExecutionFailedEventDetails

Contains details about an execution failure event.

ExecutionListItem

Contains details about an execution.

ExecutionStartedEventDetails

Contains details about the start of the execution.

ExecutionSucceededEventDetails

Contains details about the successful termination of the execution.

ExecutionTimedOutEventDetails

Contains details about the execution timeout that occurred during the execution.

HistoryEvent

Contains details about the events of an execution.

HistoryEventExecutionDataDetails

Provides details about input or output in an execution history event.

LambdaFunctionFailedEventDetails

Contains details about a lambda function that failed during an execution.

LambdaFunctionScheduleFailedEventDetails

Contains details about a failed lambda function schedule event that occurred during an

LambdaFunctionScheduledEventDetails

Contains details about a lambda function scheduled during an execution.

LambdaFunctionStartFailedEventDetails

Contains details about a lambda function that failed to start during an execution.

LambdaFunctionSucceededEventDetails

Contains details about a lambda function that successfully terminated during an

LambdaFunctionTimedOutEventDetails

Contains details about a lambda function timeout that occurred during an execution.

LogDestination

LoggingConfiguration

The LoggingConfiguration data type is used to set CloudWatch Logs

MapIterationEventDetails

Contains details about an iteration of a Map state.

MapStateStartedEventDetails

Details about a Map state that was started.

StateEnteredEventDetails

Contains details about a state entered during an execution.

StateExitedEventDetails

Contains details about an exit from a state during an execution.

StateMachineListItem

Contains details about the state machine.

Tag

Tags are key-value pairs that can be associated with Step Functions state machines and

TaskFailedEventDetails

Contains details about a task failure event.

TaskScheduledEventDetails

Contains details about a task scheduled during an execution.

TaskStartFailedEventDetails

Contains details about a task that failed to start during an execution.

TaskStartedEventDetails

Contains details about the start of a task during an execution.

TaskSubmitFailedEventDetails

Contains details about a task that failed to submit during an execution.

TaskSubmittedEventDetails

Contains details about a task submitted to a resource .

TaskSucceededEventDetails

Contains details about the successful completion of a task state.

TaskTimedOutEventDetails

Contains details about a resource timeout that occurred during an execution.

TracingConfiguration

Selects whether or not the state machine's AWS X-Ray tracing is enabled. Default is false

Enums

ExecutionStatus

HistoryEventType

LogLevel

StateMachineStatus

StateMachineType

SyncExecutionStatus