Struct aws_sdk_sfn::model::history_event::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for HistoryEvent

Implementations

impl Builder

pub fn timestamp(self, input: Instant) -> Self

The date and time the event occurred.

pub fn set_timestamp(self, input: Option<Instant>) -> Self

The date and time the event occurred.

pub fn type(self, input: HistoryEventType) -> Self

The type of the event.

pub fn set_type(self, input: Option<HistoryEventType>) -> Self

The type of the event.

pub fn id(self, input: i64) -> Self

The id of the event. Events are numbered sequentially, starting at one.

pub fn set_id(self, input: Option<i64>) -> Self

The id of the event. Events are numbered sequentially, starting at one.

pub fn previous_event_id(self, input: i64) -> Self

The id of the previous event.

pub fn set_previous_event_id(self, input: Option<i64>) -> Self

The id of the previous event.

pub fn activity_failed_event_details(
self,
input: ActivityFailedEventDetails
) -> Self

Contains details about an activity that failed during an execution.

pub fn set_activity_failed_event_details(
self,
input: Option<ActivityFailedEventDetails>
) -> Self

Contains details about an activity that failed during an execution.

pub fn activity_schedule_failed_event_details(
self,
input: ActivityScheduleFailedEventDetails
) -> Self

Contains details about an activity schedule event that failed during an execution.

pub fn set_activity_schedule_failed_event_details(
self,
input: Option<ActivityScheduleFailedEventDetails>
) -> Self

Contains details about an activity schedule event that failed during an execution.

pub fn activity_scheduled_event_details(
self,
input: ActivityScheduledEventDetails
) -> Self

Contains details about an activity scheduled during an execution.

pub fn set_activity_scheduled_event_details(
self,
input: Option<ActivityScheduledEventDetails>
) -> Self

Contains details about an activity scheduled during an execution.

pub fn activity_started_event_details(
self,
input: ActivityStartedEventDetails
) -> Self

Contains details about the start of an activity during an execution.

pub fn set_activity_started_event_details(
self,
input: Option<ActivityStartedEventDetails>
) -> Self

Contains details about the start of an activity during an execution.

pub fn activity_succeeded_event_details(
self,
input: ActivitySucceededEventDetails
) -> Self

Contains details about an activity that successfully terminated during an execution.

pub fn set_activity_succeeded_event_details(
self,
input: Option<ActivitySucceededEventDetails>
) -> Self

Contains details about an activity that successfully terminated during an execution.

pub fn activity_timed_out_event_details(
self,
input: ActivityTimedOutEventDetails
) -> Self

Contains details about an activity timeout that occurred during an execution.

pub fn set_activity_timed_out_event_details(
self,
input: Option<ActivityTimedOutEventDetails>
) -> Self

Contains details about an activity timeout that occurred during an execution.

pub fn task_failed_event_details(self, input: TaskFailedEventDetails) -> Self

Contains details about the failure of a task.

pub fn set_task_failed_event_details(
self,
input: Option<TaskFailedEventDetails>
) -> Self

Contains details about the failure of a task.

pub fn task_scheduled_event_details(
self,
input: TaskScheduledEventDetails
) -> Self

Contains details about a task that was scheduled.

pub fn set_task_scheduled_event_details(
self,
input: Option<TaskScheduledEventDetails>
) -> Self

Contains details about a task that was scheduled.

pub fn task_start_failed_event_details(
self,
input: TaskStartFailedEventDetails
) -> Self

Contains details about a task that failed to start.

pub fn set_task_start_failed_event_details(
self,
input: Option<TaskStartFailedEventDetails>
) -> Self

Contains details about a task that failed to start.

pub fn task_started_event_details(self, input: TaskStartedEventDetails) -> Self

Contains details about a task that was started.

pub fn set_task_started_event_details(
self,
input: Option<TaskStartedEventDetails>
) -> Self

Contains details about a task that was started.

pub fn task_submit_failed_event_details(
self,
input: TaskSubmitFailedEventDetails
) -> Self

Contains details about a task that where the submit failed.

pub fn set_task_submit_failed_event_details(
self,
input: Option<TaskSubmitFailedEventDetails>
) -> Self

Contains details about a task that where the submit failed.

pub fn task_submitted_event_details(
self,
input: TaskSubmittedEventDetails
) -> Self

Contains details about a submitted task.

pub fn set_task_submitted_event_details(
self,
input: Option<TaskSubmittedEventDetails>
) -> Self

Contains details about a submitted task.

pub fn task_succeeded_event_details(
self,
input: TaskSucceededEventDetails
) -> Self

Contains details about a task that succeeded.

pub fn set_task_succeeded_event_details(
self,
input: Option<TaskSucceededEventDetails>
) -> Self

Contains details about a task that succeeded.

pub fn task_timed_out_event_details(
self,
input: TaskTimedOutEventDetails
) -> Self

Contains details about a task that timed out.

pub fn set_task_timed_out_event_details(
self,
input: Option<TaskTimedOutEventDetails>
) -> Self

Contains details about a task that timed out.

pub fn execution_failed_event_details(
self,
input: ExecutionFailedEventDetails
) -> Self

Contains details about an execution failure event.

pub fn set_execution_failed_event_details(
self,
input: Option<ExecutionFailedEventDetails>
) -> Self

Contains details about an execution failure event.

pub fn execution_started_event_details(
self,
input: ExecutionStartedEventDetails
) -> Self

Contains details about the start of the execution.

pub fn set_execution_started_event_details(
self,
input: Option<ExecutionStartedEventDetails>
) -> Self

Contains details about the start of the execution.

pub fn execution_succeeded_event_details(
self,
input: ExecutionSucceededEventDetails
) -> Self

Contains details about the successful termination of the execution.

pub fn set_execution_succeeded_event_details(
self,
input: Option<ExecutionSucceededEventDetails>
) -> Self

Contains details about the successful termination of the execution.

pub fn execution_aborted_event_details(
self,
input: ExecutionAbortedEventDetails
) -> Self

Contains details about an abort of an execution.

pub fn set_execution_aborted_event_details(
self,
input: Option<ExecutionAbortedEventDetails>
) -> Self

Contains details about an abort of an execution.

pub fn execution_timed_out_event_details(
self,
input: ExecutionTimedOutEventDetails
) -> Self

Contains details about the execution timeout that occurred during the execution.

pub fn set_execution_timed_out_event_details(
self,
input: Option<ExecutionTimedOutEventDetails>
) -> Self

Contains details about the execution timeout that occurred during the execution.

pub fn map_state_started_event_details(
self,
input: MapStateStartedEventDetails
) -> Self

Contains details about Map state that was started.

pub fn set_map_state_started_event_details(
self,
input: Option<MapStateStartedEventDetails>
) -> Self

Contains details about Map state that was started.

pub fn map_iteration_started_event_details(
self,
input: MapIterationEventDetails
) -> Self

Contains details about an iteration of a Map state that was started.

pub fn set_map_iteration_started_event_details(
self,
input: Option<MapIterationEventDetails>
) -> Self

Contains details about an iteration of a Map state that was started.

pub fn map_iteration_succeeded_event_details(
self,
input: MapIterationEventDetails
) -> Self

Contains details about an iteration of a Map state that succeeded.

pub fn set_map_iteration_succeeded_event_details(
self,
input: Option<MapIterationEventDetails>
) -> Self

Contains details about an iteration of a Map state that succeeded.

pub fn map_iteration_failed_event_details(
self,
input: MapIterationEventDetails
) -> Self

Contains details about an iteration of a Map state that failed.

pub fn set_map_iteration_failed_event_details(
self,
input: Option<MapIterationEventDetails>
) -> Self

Contains details about an iteration of a Map state that failed.

pub fn map_iteration_aborted_event_details(
self,
input: MapIterationEventDetails
) -> Self

Contains details about an iteration of a Map state that was aborted.

pub fn set_map_iteration_aborted_event_details(
self,
input: Option<MapIterationEventDetails>
) -> Self

Contains details about an iteration of a Map state that was aborted.

pub fn lambda_function_failed_event_details(
self,
input: LambdaFunctionFailedEventDetails
) -> Self

Contains details about a lambda function that failed during an execution.

pub fn set_lambda_function_failed_event_details(
self,
input: Option<LambdaFunctionFailedEventDetails>
) -> Self

Contains details about a lambda function that failed during an execution.

pub fn lambda_function_schedule_failed_event_details(
self,
input: LambdaFunctionScheduleFailedEventDetails
) -> Self

Contains details about a failed lambda function schedule event that occurred during an execution.

pub fn set_lambda_function_schedule_failed_event_details(
self,
input: Option<LambdaFunctionScheduleFailedEventDetails>
) -> Self

Contains details about a failed lambda function schedule event that occurred during an execution.

pub fn lambda_function_scheduled_event_details(
self,
input: LambdaFunctionScheduledEventDetails
) -> Self

Contains details about a lambda function scheduled during an execution.

pub fn set_lambda_function_scheduled_event_details(
self,
input: Option<LambdaFunctionScheduledEventDetails>
) -> Self

Contains details about a lambda function scheduled during an execution.

pub fn lambda_function_start_failed_event_details(
self,
input: LambdaFunctionStartFailedEventDetails
) -> Self

Contains details about a lambda function that failed to start during an execution.

pub fn set_lambda_function_start_failed_event_details(
self,
input: Option<LambdaFunctionStartFailedEventDetails>
) -> Self

Contains details about a lambda function that failed to start during an execution.

pub fn lambda_function_succeeded_event_details(
self,
input: LambdaFunctionSucceededEventDetails
) -> Self

Contains details about a lambda function that terminated successfully during an execution.

pub fn set_lambda_function_succeeded_event_details(
self,
input: Option<LambdaFunctionSucceededEventDetails>
) -> Self

Contains details about a lambda function that terminated successfully during an execution.

pub fn lambda_function_timed_out_event_details(
self,
input: LambdaFunctionTimedOutEventDetails
) -> Self

Contains details about a lambda function timeout that occurred during an execution.

pub fn set_lambda_function_timed_out_event_details(
self,
input: Option<LambdaFunctionTimedOutEventDetails>
) -> Self

Contains details about a lambda function timeout that occurred during an execution.

pub fn state_entered_event_details(
self,
input: StateEnteredEventDetails
) -> Self

Contains details about a state entered during an execution.

pub fn set_state_entered_event_details(
self,
input: Option<StateEnteredEventDetails>
) -> Self

Contains details about a state entered during an execution.

pub fn state_exited_event_details(self, input: StateExitedEventDetails) -> Self

Contains details about an exit from a state during an execution.

pub fn set_state_exited_event_details(
self,
input: Option<StateExitedEventDetails>
) -> Self

Contains details about an exit from a state during an execution.

pub fn build(self) -> HistoryEvent

Consumes the builder and constructs a HistoryEvent