Module aws_sdk_sfn::operation

Structs

CreateActivity

Creates an activity. An activity is a task that you write in any programming language and

CreateStateMachine

Creates a state machine. A state machine consists of a collection of states that can do

DeleteActivity

Deletes an activity.

DeleteStateMachine

Deletes a state machine. This is an asynchronous operation: It sets the state machine's

DescribeActivity

Describes an activity.

DescribeExecution

Describes an execution.

DescribeStateMachine

Describes a state machine.

DescribeStateMachineForExecution

Describes the state machine associated with a specific execution.

GetActivityTask

Used by workers to retrieve a task (with the specified activity ARN) which has been

GetExecutionHistory

Returns the history of the specified execution as a list of events. By default, the

ListActivities

Lists the existing activities.

ListExecutions

Lists the executions of a state machine that meet the filtering criteria. Results are

ListStateMachines

Lists the existing state machines.

ListTagsForResource

List tags for a given resource.

SendTaskFailure

Used by activity workers and task states using the callback

SendTaskHeartbeat

Used by activity workers and task states using the callback

SendTaskSuccess

Used by activity workers and task states using the callback

StartExecution

Starts a state machine execution.

StartSyncExecution

Starts a Synchronous Express state machine execution.

StopExecution

Stops an execution.

TagResource

Add a tag to a Step Functions resource.

UntagResource

Remove a tag from a Step Functions resource

UpdateStateMachine

Updates an existing state machine by modifying its definition,